Generator: Move request/reply generation to extra module
